create notice model
