+++ database/models/index.js
@@ -1,5 +1,7 @@
 const { UserModel } = require("./user.model");
+const NoticesModel = require("./notices.model");
 
 module.exports = {
   UserModel,
+  NoticesModel,
 };
